// src/graphql-schema.ts
export type ID = string
export type GitObjectID = string
export type DateTime = string
// ...
export interface IPageInfo {
    __typename: 'PageInfo'
    endCursor: string | null
    hasNextPage: boolean
}
// ...
export interface IUserConnection {
    __typename: 'UserConnection'
    nodes: IUser[]
    totalCount: number
    pageInfo: IPageInfo
}
// ...
export interface IPermissionsInfo {
    __typename: 'PermissionsInfo'
    permissions: RepositoryPermission[]
    syncedAt: DateTime | null
    updatedAt: DateTime
}

export enum RepositoryPermission {
    READ = 'READ',
}

export interface IRepositoryComparisonConnection {
    __typename: 'RepositoryComparisonConnection'
    nodes: IRepositoryComparison[]
    totalCount: number
    pageInfo: IPageInfo
}

export interface IRepositoryComparison {
    __typename: 'RepositoryComparison'
    baseRepository: IRepository
    headRepository: IRepository
    range: IGitRevisionRange
    commits: IGitCommitConnection
    fileDiffs: IFileDiffConnection
}

export interface ICommitsOnRepositoryComparisonArguments {
    first?: number | null
}

export interface IFileDiffsOnRepositoryComparisonArguments {
    first?: number | null
    after?: string | null
}

export interface IRepository {
    __typename: 'Repository'
    id: ID
    name: string

    /**
     * @deprecated "Use name."
     */
    uri: string
    description: string
    language: string
    createdAt: DateTime
    updatedAt: DateTime | null
    commit: IGitCommit | null
    mirrorInfo: IMirrorRepositoryInfo
    externalRepository: IExternalRepository
    isFork: boolean
    isArchived: boolean
    isPrivate: boolean
    externalServices: IExternalServiceConnection

    /**
     * @deprecated "use Repository.mirrorInfo.cloneInProgress instead"
     */
    cloneInProgress: boolean
    textSearchIndex: IRepositoryTextSearchIndex | null
    url: string
    externalURLs: IExternalLink[]
    defaultBranch: IGitRef | null
    gitRefs: IGitRefConnection
    branches: IGitRefConnection
    tags: IGitRefConnection
    comparison: IRepositoryComparison
    contributors: IRepositoryContributorConnection
    viewerCanAdminister: boolean
    icon: string
    label: IMarkdown
    detail: IMarkdown
    matches: ISearchResultMatch[]
    lsifUploads: ILSIFUploadConnection
    lsifIndexes: ILSIFIndexConnection
    authorizedUsers: IUserConnection
    permissionsInfo: IPermissionsInfo | null
}
// ...
export interface IMarkdown {
    __typename: 'Markdown'
    text: string
    html: string
}

export interface ISearchResultMatch {
    __typename: 'SearchResultMatch'
    url: string
    body: IMarkdown
    highlights: IHighlight[]
}

export interface IHighlight {
    __typename: 'Highlight'
    line: number
    character: number
    length: number
}

export interface IGitCommit {
    __typename: 'GitCommit'
    id: ID
    repository: IRepository
    oid: GitObjectID
    abbreviatedOID: string
    author: ISignature
    committer: ISignature | null
    message: string
    subject: string
    body: string | null
    parents: IGitCommit[]
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    tree: IGitTree | null
    blob: IGitBlob | null
    file: File2 | null
    languages: string[]
    languageStatistics: ILanguageStatistics[]
    ancestors: IGitCommitConnection
    behindAhead: IBehindAheadCounts
    symbols: ISymbolConnection
}
// ...
export interface ISignature {
    __typename: 'Signature'
    person: IPerson
    date: string
}

export interface IPerson {
    __typename: 'Person'
    name: string
    email: string
    displayName: string
    avatarURL: string
    user: IUser | null
}

export interface IExternalLink {
    __typename: 'ExternalLink'
    url: string
    serviceType: string | null
}

export interface IGitTree {
    __typename: 'GitTree'
    path: string
    isRoot: boolean
    name: string
    isDirectory: boolean
    commit: IGitCommit
    repository: IRepository
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    rawZipArchiveURL: string
    submodule: ISubmodule | null
    directories: IGitTree[]
    files: IFile[]
    entries: TreeEntry[]
    symbols: ISymbolConnection
    isSingleChild: boolean
    lsif: TreeEntryLSIFData | null
}
// ...
export type TreeEntry = IGitTree | IGitBlob
// ...
export interface ISymbolConnection {
    __typename: 'SymbolConnection'
    nodes: ISymbol[]
    pageInfo: IPageInfo
}

export interface ISymbol {
    __typename: 'Symbol'
    name: string
    containerName: string | null
    kind: SymbolKind
    language: string
    location: ILocation
    url: string
    canonicalURL: string
    fileLocal: boolean
}
// ...
export interface IGitBlob {
    __typename: 'GitBlob'
    path: string
    name: string
    isDirectory: boolean
    content: string
    byteSize: number
    binary: boolean
    richHTML: string
    commit: IGitCommit
    repository: IRepository
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    blame: IHunk[]
    highlight: IHighlightedFile
    submodule: ISubmodule | null
    symbols: ISymbolConnection
    isSingleChild: boolean
    lsif: IGitBlobLSIFData | null
}
// ...
export type File2 = IGitBlob | IVirtualFile
// ...
export interface IHighlightedFile {
    __typename: 'HighlightedFile'
    aborted: boolean
    html: string
}

export interface IHunk {
    __typename: 'Hunk'
    startLine: number
    endLine: number
    startByte: number
    endByte: number
    rev: string
    author: ISignature
    message: string
    commit: IGitCommit
}

export interface ISubmodule {
    __typename: 'Submodule'
    url: string
    commit: string
    path: string
}

export interface IGitBlobLSIFData {
    __typename: 'GitBlobLSIFData'
    definitions: ILocationConnection | null
    references: ILocationConnection | null
    hover: IHover | null
    diagnostics: IDiagnosticConnection
}
// ...
export type TreeEntryLSIFData = IGitBlobLSIFData
// ...
export interface IFile {
    __typename: 'File'
    path: string
    name: string
    isDirectory: boolean
    url: string
    repository: IRepository
}

export interface ILanguageStatistics {
    __typename: 'LanguageStatistics'
    name: string
    totalBytes: number
    totalLines: number
}

export interface IGitCommitConnection {
    __typename: 'GitCommitConnection'
    nodes: IGitCommit[]
    totalCount: number | null
    pageInfo: IPageInfo
}

export interface IBehindAheadCounts {
    __typename: 'BehindAheadCounts'
    behind: number
    ahead: number
}

export interface IMirrorRepositoryInfo {
    __typename: 'MirrorRepositoryInfo'
    remoteURL: string
    cloneInProgress: boolean
    cloneProgress: string | null
    cloned: boolean
    updatedAt: DateTime | null
    updateSchedule: IUpdateSchedule | null
    updateQueue: IUpdateQueue | null
}

export interface IUpdateSchedule {
    __typename: 'UpdateSchedule'
    intervalSeconds: number
    due: DateTime
    index: number
    total: number
}

export interface IUpdateQueue {
    __typename: 'UpdateQueue'
    index: number
    updating: boolean
    total: number
}

export interface IExternalRepository {
    __typename: 'ExternalRepository'
    id: string
    serviceType: string
    serviceID: string
}

export interface IExternalServiceConnection {
    __typename: 'ExternalServiceConnection'
    nodes: IExternalService[]
    totalCount: number
    pageInfo: IPageInfo
}
// ...
export interface IRepositoryTextSearchIndex {
    __typename: 'RepositoryTextSearchIndex'
    repository: IRepository
    status: IRepositoryTextSearchIndexStatus | null
    refs: IRepositoryTextSearchIndexedRef[]
}
// ...
export interface IGitRef {
    __typename: 'GitRef'
    id: ID
    name: string
    abbrevName: string
    displayName: string
    prefix: string
    type: GitRefType
    target: IGitObject
    repository: IRepository
    url: string
}
// ...
export interface IGitRefConnection {
    __typename: 'GitRefConnection'
    nodes: IGitRef[]
    totalCount: number
    pageInfo: IPageInfo
}

export interface IRepositoryContributorConnection {
    __typename: 'RepositoryContributorConnection'
    nodes: IRepositoryContributor[]
    totalCount: number
    pageInfo: IPageInfo
}
// ...
export interface ILSIFUploadConnection {
    __typename: 'LSIFUploadConnection'
    nodes: ILSIFUpload[]
    totalCount: number | null
    pageInfo: IPageInfo
}
// ...
export interface ILSIFIndexConnection {
    __typename: 'LSIFIndexConnection'
    nodes: ILSIFIndex[]
    totalCount: number | null
    pageInfo: IPageInfo
}
// ...
export interface IVirtualFile {
    __typename: 'VirtualFile'
    path: string
    name: string
    isDirectory: boolean
    content: string
    byteSize: number
    binary: boolean
    richHTML: string
    url: string
    canonicalURL: string
    externalURLs: IExternalLink[]
    highlight: IHighlightedFile
}